requirements.txt
streamlit
huggingface
transformers
google.generativeai
torch
pillow
python-dotenv
requests
langchain
langchain_community
fpdf